Add time import and DatabaseConfig struct to db.go; remove models.go
 - Import the "time" package in db.go for time.Duration usage.
 - Introduce DatabaseConfig struct in db.go with connection parameters.
 - Delete models.go, moving DatabaseConfig to db.go.
